// src/app/modules/academicSemester/academicSemester.interface.ts
export type IMonth =
  | 'January'
  | 'February'
  | 'March'
  | 'April'
  | 'May'
  | 'June'
  | 'July'
  | 'August'
  | 'September'
  | 'October'
  | 'November'
  | 'December'

export type IAcademicSemester = {
  title: 'Autumn' | 'Summer' | 'Fall'
  year: number
  code: '01' | '02' | '03'
  startMonth: IMonth
  endMonth: IMonth
}